+++ src/contexts/index.ts
@@ -2,7 +2,8 @@ export {
   type Settings,
   SettingsContext,
   type SettingsContextProps,
-  Theme,
+  THEMES,
+  type Theme,
 } from "./SettingsContext";
 export { SettingsProvider, useSettingsContext } from "./SettingsProvider";
 export { VersionContext, type VersionContextProps } from "./VersionContext";